pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools> 61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "text2story"
version = "1.2.11"
readme = "README.md"
dependencies = [
	"scipy",
	"numpy < 1.24",
   	"pandas < 2.0",
        "torchaudio < 0.12",
        "torchdata <= 0.3.0",
        "torchtext <= 0.12",
   	"nltk",
        "datasets<=2.10.1",
   	"overrides==3.1.0",
	"cached-path==1.1.2",
	"joblib",
   	"requests< 2.28.0",
	"allennlp==2.9.3",
   	"allennlp-models==2.9.3",
	"spacy==3.2.0",
	"pyspark",
	"spark-nlp",
	"elementpath",
	"sklearn-crfsuite",
	"plantuml",
	"pdflatex",
	"pdf2image",
	"pytest<=6.1.0",
	"py-heideltime",
        "pyarrow<10.0",
	"gdown",
        "inflect<6.0.0",
        "google-auth < 2.18.0"
        
]

[project.urls]
Home = "https://github.com/LIAAD/Text2StoryPackage"


